10lb/requirements.txt
flask==2.0.1
pillow==8.3.1
cryptography==35.0.0